+++ ui/mods/com.pa.quitch.gwaioverhaul/gw_play/card_units.js
@@ -1398,5 +1398,13 @@ define({
       id: "gwaio_upgrade_orbitalfactory",
       units: ["/pa/units/orbital/orbital_factory/orbital_factory.json"],
     },
+    {
+      id: "gwaio_enable_planetaryradar",
+      units: ["/pa/units/orbital/deep_space_radar/deep_space_radar.json"],
+    },
+    {
+      id: "gwaio_upgrade_planetaryradar",
+      units: ["/pa/units/orbital/deep_space_radar/deep_space_radar.json"],
+    },
   ],
 });